fix: Resolve clippy warnings and update CI to check only HAL code
Fixed all clippy issues causing CI failures.

Code Fixes (efr32mg24-hal/src/usart/mod.rs:123):
- Removed unnecessary cast in USART baud rate calculation
- Changed (clkdiv & 0xFFFFF) as u32 to (clkdiv & 0xFFFFF)

CI Workflow Updates (.github/workflows/ci.yml):
- Added --no-deps flag to HAL clippy check (line 62)
- Now checks only HAL code, not dependencies (PAC)
- PAC warnings no longer cascade to HAL clippy check

Pre-Push Script Updates (.claude/skills/ci-cd-tester/scripts/pre-push-check.sh):
- Updated HAL clippy check to use --no-deps flag
- Simplified logic and improved reliability

This should resolve all GitHub Actions CI failures.

+++ efr32mg24-hal/src/usart/mod.rs
@@ -120,7 +120,7 @@ impl Usart0 {
         let clkdiv = (256 * pclk / (oversample * config.baudrate)) - 256;
         usart
             .clkdiv()
-            .write(|w| unsafe { w.div().bits((clkdiv & 0xFFFFF) as u32) });
+            .write(|w| unsafe { w.div().bits(clkdiv & 0xFFFFF) });
 
         // Enable TX and RX
         usart.cmd().write(|w| {
